refactor: move distance sensor to new package, rename DistanceSensorInterface to DistanceSensor
