+++ src/main/java/frc/robot/lib/sensor/distance/DistanceSensor.java
@@ -1,6 +1,6 @@
-package frc.robot.lib;
+package frc.robot.lib.sensor.distance;
 
-public interface DistanceSensorInterface {
+public interface DistanceSensor {
   public double getTargetDistance();
 
   public boolean isGetTarget();
